# [framework] Speed up feedthrough calculation during diagram building