--- a/systems/framework/system_base.cc
+++ b/systems/framework/system_base.cc
@@ -287,6 +287,21 @@ CacheEntry& SystemBase::DeclareCacheEntryWithKnownTicket(
   return new_entry;
 }
 
+bool SystemBase::IsObviouslyNotInputDependent(
+    DependencyTicket dependency_ticket) const {
+  return (dependency_ticket <= internal::kAllSourcesExceptInputPortsTicket) ||
+         std::any_of(discrete_state_tickets_.begin(),
+                     discrete_state_tickets_.end(),
+                     [&dependency_ticket](const auto& info) {
+                       return info.ticket == dependency_ticket;
+                     }) ||
+         std::any_of(abstract_state_tickets_.begin(),
+                     abstract_state_tickets_.end(),
+                     [&dependency_ticket](const auto& info) {
+                       return info.ticket == dependency_ticket;
+                     });
+}
+
 void SystemBase::InitializeContextBase(ContextBase* context_ptr) const {
   DRAKE_DEMAND(context_ptr != nullptr);
   ContextBase& context = *context_ptr;