When possible, we now avoid calling AllocateContext when answering direct feedthrough queries.