@jwnimmer-tri 2024-06-27
> Is the Context allocation a significant hit in practice ...

For instance, SceneGraph does a crazy amount of stuff when setting up a Context.  However, I just realized that most of that work seems to be for setting up a _default_ context, instead of merely _allocating_ a context, so maybe this doesn't change much?

Basically I was curious to get your thoughts about the most likely hot spots or penalties, to see if we should try to dig a bit deeper and look at some stats.

> ... and do these short circuits avoid it?

For SceneGraph at least, I believe so (but I haven't tested it) -- all of the SG output ports declare `all_input_ports_ticket()` so this function is really fast and boring now.

I wanted to get a second opinion before I dug in too deep.