@jwnimmer-tri 2024-06-30
Benchmarks from `bazel run //systems/benchmarking:framework_experiment -- --output_dir=try1 --sleep=1 -- --benchmark_filter=DiagramBuild`:

On master:
```
----------------------------------------------------------------------
Benchmark                  Time             CPU    Allocs   Iterations
----------------------------------------------------------------------
DiagramBuild/3/0       0.196 ms        0.197 ms      2.7k         3546
DiagramBuild/30/0       9.66 ms         9.66 ms    166.4k           73
DiagramBuild/3/1       0.455 ms        0.454 ms     13.9k         1553
DiagramBuild/3/2        1.06 ms         1.06 ms     54.2k          672
```

On this PR:
```
DiagramBuild/3/0       0.228 ms        0.228 ms      2.7k         3006
DiagramBuild/30/0       11.2 ms         11.2 ms    166.4k           62
DiagramBuild/3/1       0.523 ms        0.523 ms     13.8k         1338
DiagramBuild/3/2        1.23 ms         1.23 ms     53.5k          588
```

On master after changing Adder to declare output port dependencies:
```
DiagramBuild/3/0       0.063 ms        0.063 ms      1.5k        10993
DiagramBuild/30/0       1.71 ms         1.71 ms     31.6k          406
DiagramBuild/3/1       0.123 ms        0.123 ms      6.8k         5722
DiagramBuild/3/2       0.287 ms        0.287 ms     25.9k         2411
```

On this PR after changing Adder to declare output port dependencies:
```
DiagramBuild/3/0       0.019 ms        0.019 ms   801.125        37326
DiagramBuild/30/0      0.928 ms        0.928 ms     14.1k          756
DiagramBuild/3/1       0.036 ms        0.036 ms      3.1k        19451
DiagramBuild/3/2       0.083 ms        0.082 ms     10.4k         8504
```