@sherm1 2024-06-30
Looks great except for the "801.125" allocations in this row:
> DiagramBuild/3/0       0.019 ms        0.019 ms   801.125        37326

Is that spurious?