@jwnimmer-tri 2024-06-30
Other rows have `k` for "thousand", so it's a move in a good direction.