@jwnimmer-tri 2024-07-31
Users report 10% speed-up with this patch.  I'm going to work on moving it forward.